refactor: move bound cron execution out of gateway
